Refactor runner package.
